# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Minha API",
    "version": "1.0.0",
    "description": "Documentação da API"
  },
  "servers": [
    {
      "url": "http://localhost:8888",
      "description": "Servidor Local"
    }
  ],
  "paths": {
    "/companies": {
      "get": {
        "summary": "Retorna a lista de empresas",
        "tags": [
          "Company"
        ],
        "responses": {
          "200": {
            "description": "Lista de empresas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "example": 1
                      },
                      "name": {
                        "type": "string",
                        "example": "Hakuna Matata Corp"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Cria uma nova empresa",
        "tags": [
          "Company"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Hakuna Matata Corp"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Empresa criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": "Empresa Hakuna Matata Corp criada com sucesso"
                }
              }
            }
          }
        }
      }
    },
    "/companies/{id}": {
      "get": {
        "summary": "Retorna os detalhes de uma empresa pelo ID",
        "tags": [
          "Company"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID da empresa"
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhes da empresa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "name": {
                      "type": "string",
                      "example": "Hakuna Matata Corp"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Empresa não encontrada"
          }
        }
      },
      "put": {
        "summary": "Atualiza os detalhes de uma empresa pelo ID",
        "tags": [
          "Company"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID da empresa"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Hakuna Matata Corp Updated"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Empresa atualizada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": "Empresa com ID 1 atualizada com sucesso"
                }
              }
            }
          },
          "404": {
            "description": "Empresa não encontrada"
          }
        }
      },
      "delete": {
        "summary": "Deleta uma empresa pelo ID",
        "tags": [
          "Company"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID da empresa"
          }
        ],
        "responses": {
          "200": {
            "description": "Empresa deletada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": "Empresa com ID 1 deletada com sucesso"
                }
              }
            }
          }
        }
      }
    },
    "/products": {
      "get": {
        "summary": "Retorna a lista de produtos",
        "tags": [
          "Product"
        ],
        "responses": {
          "200": {
            "description": "Lista de produtos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "example": 1
                      },
                      "name": {
                        "type": "string",
                        "example": "Estrela do Norte"
                      },
                      "price": {
                        "type": "number",
                        "example": 19.99
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Cria um novo produto",
        "tags": [
          "Product"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Estrela do Norte"
                  },
                  "price": {
                    "type": "number",
                    "example": 19.99
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Produto criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": "Produto Estrela do Norte criado com sucesso"
                }
              }
            }
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "summary": "Retorna os detalhes de um produto específico",
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID do produto"
          }
        ],
        "responses": {
          "200": {
            "description": "Detalhes do produto",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "name": {
                      "type": "string",
                      "example": "Estrela do Norte"
                    },
                    "price": {
                      "type": "number",
                      "example": 19.99
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Produto não encontrado"
          }
        }
      },
      "put": {
        "summary": "Atualiza um produto existente",
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID do produto"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Estrela do Norte Atualizado"
                  },
                  "price": {
                    "type": "number",
                    "example": 29.99
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Produto atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": "Produto com ID 1 atualizado com sucesso"
                }
              }
            }
          },
          "404": {
            "description": "Produto não encontrado"
          }
        }
      },
      "delete": {
        "summary": "Deleta um produto existente",
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID do produto"
          }
        ],
        "responses": {
          "200": {
            "description": "Produto deletado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": "Produto com ID 1 deletado com sucesso"
                }
              }
            }
          },
          "404": {
            "description": "Produto não encontrado"
          }
        }
      }
    }
  },
  "components": {},
  "tags": [
    {
      "name": "Company",
      "description": "API para gerenciamento de empresas"
    }
  ]
}